refactor: rename `config_set` file to `config_module`
